+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `ExperimentSpec.with_description()`.
+- Added `TaskSpec.with_image()`, `.with_result()`, `.with_context()`, `.with_name()`, `.with_command()`, and `.with_arguments()`.
+
+### Changed
+
+- Renamed `delete_results_dataset` param to `delete_results_datasets` in `Beaker.experiment.delete()`.
+- Renamed `TaskSpec.with_data()` to `TaskSpec.with_dataset()`.
+
+### Fixed
+
+- Fixed bug with `Beaker.experiment.delete()` that would lead to a `ValueError` being raised when the experiment to delete has multiple tasks.
+
 ## [v0.14.1](https://github.com/allenai/beaker-py/releases/tag/v0.14.1) - 2022-05-05
 
 ### Changed