+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v0.15.1](https://github.com/allenai/beaker-py/releases/tag/v0.15.1) - 2022-05-06
+
+### Added
+
+- Docker images released with major and major+minor version tags in addition to full version tag.
+
 ## [v0.15.0](https://github.com/allenai/beaker-py/releases/tag/v0.15.0) - 2022-05-06
 
 ### Added